server/requirements.txt
flask==3.0.2
pandas==2.2.1
pdfplumber==0.10.3
python-pptx==0.6.23
pytest==8.0.2
pyyaml==6.0
lark==1.2.2